Add `ORM\Query\SelectQuery`.
This makes the namespace and classname for SELECT query consistent with
other query types. `ORM\Query` is retained as alias for easier upgrading.
