+++ src/ORM/Association/HasMany.php
@@ -24,7 +24,7 @@ use Cake\Datasource\EntityInterface;
 use Cake\Datasource\InvalidPropertyInterface;
 use Cake\ORM\Association;
 use Cake\ORM\Association\Loader\SelectLoader;
-use Cake\ORM\Query;
+use Cake\ORM\Query\SelectQuery;
 use Cake\ORM\Table;
 use Closure;
 use InvalidArgumentException;
@@ -49,7 +49,7 @@ class HasMany extends Association
      *
      * @var string
      */
-    protected string $_joinType = Query::JOIN_TYPE_INNER;
+    protected string $_joinType = SelectQuery::JOIN_TYPE_INNER;
 
     /**
      * The strategy name to be used to fetch associated records.
