+++ src/ORM/BehaviorRegistry.php
@@ -22,6 +22,7 @@ use Cake\Core\ObjectRegistry;
 use Cake\Event\EventDispatcherInterface;
 use Cake\Event\EventDispatcherTrait;
 use Cake\ORM\Exception\MissingBehaviorException;
+use Cake\ORM\Query\SelectQuery;
 use LogicException;
 
 /**
@@ -265,10 +266,10 @@ class BehaviorRegistry extends ObjectRegistry implements EventDispatcherInterfac
      *
      * @param string $type The finder type to invoke.
      * @param array $args The arguments you want to invoke the method with.
-     * @return \Cake\ORM\Query The return value depends on the underlying behavior method.
+     * @return \Cake\ORM\Query\SelectQuery The return value depends on the underlying behavior method.
      * @throws \BadMethodCallException When the method is unknown.
      */
-    public function callFinder(string $type, array $args = []): Query
+    public function callFinder(string $type, array $args = []): SelectQuery
     {
         $type = strtolower($type);
