+++ src/ORM/LazyEagerLoader.php
@@ -21,6 +21,7 @@ use Cake\Collection\CollectionInterface;
 use Cake\Database\Expression\QueryExpression;
 use Cake\Database\Expression\TupleComparison;
 use Cake\Datasource\EntityInterface;
+use Cake\ORM\Query\SelectQuery;
 
 /**
  * Contains methods that are capable of injecting eagerly loaded associations into
@@ -68,9 +69,9 @@ class LazyEagerLoader
      * @param \Cake\Collection\CollectionInterface $objects The original entities
      * @param array $contain The associations to be loaded
      * @param \Cake\ORM\Table $source The table to use for fetching the top level entities
-     * @return \Cake\ORM\Query
+     * @return \Cake\ORM\Query\SelectQuery
      */
-    protected function _getQuery(CollectionInterface $objects, array $contain, Table $source): Query
+    protected function _getQuery(CollectionInterface $objects, array $contain, Table $source): SelectQuery
     {
         $primaryKey = $source->getPrimaryKey();
         $method = is_string($primaryKey) ? 'get' : 'extract';
@@ -80,7 +81,7 @@ class LazyEagerLoader
         $query = $source
             ->find()
             ->select((array)$primaryKey)
-            ->where(function (QueryExpression $exp, Query $q) use ($primaryKey, $keys, $source) {
+            ->where(function (QueryExpression $exp, SelectQuery $q) use ($primaryKey, $keys, $source) {
                 if (is_array($primaryKey) && count($primaryKey) === 1) {
                     $primaryKey = current($primaryKey);
                 }
@@ -131,13 +132,17 @@ class LazyEagerLoader
      * entities.
      *
      * @param iterable<\Cake\Datasource\EntityInterface> $objects The original list of entities
-     * @param \Cake\ORM\Query $results The loaded results
+     * @param \Cake\ORM\Query\SelectQuery $results The loaded results
      * @param array<string> $associations The top level associations that were loaded
      * @param \Cake\ORM\Table $source The table where the entities came from
      * @return array<\Cake\Datasource\EntityInterface>
      */
-    protected function _injectResults(iterable $objects, Query $results, array $associations, Table $source): array
-    {
+    protected function _injectResults(
+        iterable $objects,
+        SelectQuery $results,
+        array $associations,
+        Table $source
+    ): array {
         $injected = [];
         $properties = $this->_getPropertyMap($source, $associations);
         $primaryKey = (array)$source->getPrimaryKey();
