+++ tests/TestCase/ORM/EagerLoaderTest.php
@@ -21,7 +21,7 @@ use Cake\Database\Expression\QueryExpression;
 use Cake\Database\TypeMap;
 use Cake\Datasource\ConnectionManager;
 use Cake\ORM\EagerLoader;
-use Cake\ORM\Query;
+use Cake\ORM\Query\SelectQuery;
 use Cake\TestSuite\TestCase;
 use InvalidArgumentException;
 
@@ -189,7 +189,7 @@ class EagerLoaderTest extends TestCase
             ],
         ];
 
-        $query = $this->getMockBuilder('Cake\ORM\Query')
+        $query = $this->getMockBuilder(SelectQuery::class)
             ->onlyMethods(['join'])
             ->setConstructorArgs([$this->table])
             ->getMock();
@@ -410,7 +410,7 @@ class EagerLoaderTest extends TestCase
         ]);
         $builder = $loader->getContain()['clients']['queryBuilder'];
         $table = $this->getTableLocator()->get('foo');
-        $query = new Query($table);
+        $query = new SelectQuery($table);
         $query = $builder($query);
         $this->assertEquals(['a', 'b'], $query->clause('select'));
     }
@@ -430,7 +430,7 @@ class EagerLoaderTest extends TestCase
         ];
 
         $table = $this->getTableLocator()->get('foo');
-        $query = new Query($table);
+        $query = new SelectQuery($table);
         $loader = new EagerLoader();
         $loader->contain($contains);
         $query->select('foo.id');
@@ -454,7 +454,7 @@ class EagerLoaderTest extends TestCase
     {
         $contains = ['clients' => ['orders']];
 
-        $query = new Query($this->table);
+        $query = new SelectQuery($this->table);
         $query->select()->contain($contains)->sql();
         $select = $query->clause('select');
         $expected = [
@@ -467,7 +467,7 @@ class EagerLoaderTest extends TestCase
         $this->assertEquals($expected, $select);
 
         $contains['clients']['fields'] = ['name'];
-        $query = new Query($this->table);
+        $query = new SelectQuery($this->table);
         $query->select('foo.id')->contain($contains)->sql();
         $select = $query->clause('select');
         $expected = ['foo__id' => 'foo.id', 'clients__name' => 'clients.name'];
@@ -476,7 +476,7 @@ class EagerLoaderTest extends TestCase
 
         $contains['clients']['fields'] = [];
         $contains['clients']['orders']['fields'] = false;
-        $query = new Query($this->table);
+        $query = new SelectQuery($this->table);
         $query->select()->contain($contains)->sql();
         $select = $query->clause('select');
         $expected = [
@@ -507,7 +507,7 @@ class EagerLoaderTest extends TestCase
             ],
         ];
 
-        $query = $this->getMockBuilder('Cake\ORM\Query')
+        $query = $this->getMockBuilder(SelectQuery::class)
             ->onlyMethods(['join'])
             ->setConstructorArgs([$this->table])
             ->getMock();
